refactor(tests): reorganize fixtures for hybrid retriever
